// apps/humayro/lib/api/model/notificationDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */
import type { NotificationDTOType } from './notificationDTOType';

export interface NotificationDTO {
  id?: number;
  type?: NotificationDTOType;
  title?: string;
  body?: string;
  payloadJson?: string;
  readStatus?: boolean;
  createdAt?: string;
}